requirements.txt
absl-py==0.12.0
alabaster==0.7.12
antlr4-python3-runtime==4.8
anyio==2.2.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==19.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
boto3==1.9.246
botocore==1.12.246
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1613246939562/work
chardet @ file:///tmp/build/80754af9/chardet_1607706768982/work
click==7.1.2
coloredlogs==14.0
cryptography @ file:///tmp/build/80754af9/cryptography_1616769182610/work
cycler==0.10.0
decorator==5.0.5
defusedxml==0.7.1
docformatter==1.3
docutils==0.15
emoji==0.5.4
entrypoints==0.3
faiss-gpu==1.7.0
filelock==3.0.12
flake8==3.7.8
flake8-bugbear==19.8.0
fvcore==0.1.1.post20200716
gitdb==4.0.7
gitdb2==2.0.5
GitPython==3.0.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
grpcio==1.36.1
humanfriendly==9.1
hydra-core==1.0.3
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imagesize==1.2.0
importlib-metadata==3.10.0
importlib-resources==5.1.2
iniconfig==1.1.1
iopath==0.1.8
ipykernel==5.5.3
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==0.14.1
json5==0.9.5
jsonlines==1.2.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
kiwisolver==1.3.1
Markdown==2.6.11
markdown-it-py==0.5.8
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.7.0
myst-parser==0.12.10
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.4.5
notebook==6.3.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
oauthlib==3.1.0
omegaconf==2.0.2
packaging==20.9
pandas==1.1.1
pandocfilters==1.4.3
# Editable install with no version control (parlai==0.10.0)
-e ParlAI
parso==0.8.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
portalocker==2.3.0
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.15.7
ptyprocess==0.7.0
py==1.10.0
py-gfm==0.1.4
py-rouge==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyflakes==2.1.1
Pygments==2.8.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
pytest==5.3.2
pytest-datadir==1.3.1
pytest-regressions==2.1.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
pyzmq==18.1.0
regex==2020.1.8
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-mock==1.7.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.2.1
sacremoses==0.0.45
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
sh==1.12.14
six @ file:///tmp/build/80754af9/six_1605205313296/work
smmap==4.0.0
smmap2==3.0.1
sniffio==1.2.0
snowballstemmer==2.1.0
Sphinx==2.2.0
sphinx-autodoc-typehints==1.10.3
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
subword-nmt==0.3.7
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
torch==1.7.1
torchtext==0.8.1
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1615925068909/work
traitlets==5.0.5
transformers==4.5.1
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
Unidecode==1.2.0
untokenize==0.1.1
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.56.0
websocket-server==0.4
Werkzeug==1.0.1
yacs==0.1.8
zipp==3.4.1
